post

+++ app/Http/Controllers/TasksController.php
@@ -17,11 +17,19 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
-        
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        $data = [];
+        if(\Auth::check()){
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at','desc')->get();
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            return view('tasks.index', $data);
+        }else{
+            return view('tasks.welcome');
+        }
     }
 
     /**
@@ -50,12 +58,12 @@ class TasksController extends Controller
             'status' => 'required|max:10','content' => 'required|max:191',
         ]);
         
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        $request->user()->tasks()->create([
+            'status' => $request->status,'content' => $request->content,
+        ]);
         
-        return redirect('/');
+        return redirect()->back();
+        
     }
 
     /**
@@ -117,9 +125,18 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
+        $task = \App\Task::find($id);
         
+        if(\Auth::id() === $task->user_id){
+            $task->delete();
+        }
+            
         return redirect('/');
+    
+    //以前の部分
+    //$task = Task::find($id);
+    //$task->delete();
+    
+    //return redirect('/');
     }
 }
